@@ -1,45 +1,47 @@
-import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
-  NavLink
-} from 'react-router-dom';
+import { BrowserRouter } from "react-router-dom";
+import { Routes, Route, NavLink, Navigate } from "react-router-dom";
+import ShoppingPage from "../02-component-patterns/pages/ShoppingPage";
+// Routes
 
-import logo from '../logo.svg';
+// Components
 
-export const Navigation = () => {
+// Styles
+import logo from "../logo.svg";
+// App
+const Navigation = () => {
   return (
-    <Router>
+    <BrowserRouter>
       <div className="main-layout">
         <nav>
-            <img src={ logo } alt="React Logo" />
+          <img src={logo} alt="React Logo" />
           <ul>
             <li>
-              <NavLink to="/" activeClassName="nav-active" exact>Home</NavLink>
+              <NavLink to="/" className={({ isActive }) => (isActive ? "nav-active" : "")}>
+                Shopping
+              </NavLink>
             </li>
             <li>
-              <NavLink to="/about" activeClassName="nav-active" exact>About</NavLink>
+              <NavLink to="/about" className={({ isActive }) => (isActive ? "nav-active" : "")}>
+                About
+              </NavLink>
             </li>
             <li>
-              <NavLink to="/users" activeClassName="nav-active" exact>Users</NavLink>
+              <NavLink to="/users" className={({ isActive }) => (isActive ? "nav-active" : "")}>
+                Users
+              </NavLink>
             </li>
           </ul>
         </nav>
+        <Routes>
+          <Route path="/" element={<ShoppingPage />} />
+          <Route path="/about" element={<h1>Users</h1>} />
+          <Route path="/users" element={<h1>Home</h1>} />
 
-        {/* A <Switch> looks through its children <Route>s and
-            renders the first one that matches the current URL. */}
-        <Switch>
-          <Route path="/about">
-            <h1>About</h1>
-          </Route>
-          <Route path="/users">
-            <h1>Users</h1>
-          </Route>
-          <Route path="/">
-            <h1>Home</h1>
-          </Route>
-        </Switch>
+          <Route path="/*" element={<Navigate to={"/"} />} />
+        </Routes>
       </div>
-    </Router>
+    </BrowserRouter>
   );
-}
+};
+
+export default Navigation;
